@@ -205,7 +205,7 @@ internal class Layer
 
     internal virtual Stream OpenBackingFile() => File.OpenRead(BackingFile);
 
-    private readonly static char[] PathSeparators = new char[] { Path.DirectorySeparatorChar, Path.AltDirectorySeparatorChar };
+    private static readonly char[] PathSeparators = new char[] { Path.DirectorySeparatorChar, Path.AltDirectorySeparatorChar };
 
     /// <summary>
     /// A stream capable of computing the hash digest of raw uncompressed data while also compressing it.
